- change some nginx related files. - replace traefik with nginx. - replace jest with mocha.
