+++ k8s/app-dep.yml
@@ -6,7 +6,7 @@ spec:
   selector:
     matchLabels:
       app: app
-  replicas: 1
+  replicas: 2
   template:
     metadata:
       labels:
