+++ k8s/nginx-ingress.yml
@@ -1,0 +1,18 @@
+apiVersion: extensions/v1beta1
+kind: Ingress
+metadata:
+  name: nginx-ingress
+  annotations:
+    nginx.ingress.kubernetes.io/rewrite-target: /
+spec:
+  backend:
+    serviceName: default-http-backend
+    servicePort: 80
+  rules:
+  - host: 
+    http:
+      paths:
+        - path: /
+          backend:
+            serviceName: app-service
+            servicePort: 8080
