+++ makefile
@@ -1,5 +1,7 @@
 # docker compose
 up:
+	eval $$(minikube docker-env -u) && \
+	eval $$(docker-machine env -u) && \
 	docker-compose up -d --build
 
 ssh:
@@ -12,8 +14,19 @@ clean-db:
 	rm -rf pgdata-development pgdata-test
 
 # kubernetes
+minikube:
+	minikube addons enable ingress && \
+	minikube start && \
+	eval $$(minikube docker-env)
+
+minikube-delete:
+	make kube-delete; \
+	eval $$(minikube docker-env -u) && \
+	minikube stop && \
+	minikube delete
+
 kube:
-	@eval $$(minikube docker-env) && \
+	eval $$(minikube docker-env) && \
 	make kube-delete; \
 	make kube-create
 
@@ -21,32 +34,8 @@ kube-migrate:
 	kubectl exec $(shell kubectl get pods -l app=job -o name --field-selector=status.phase==Running | sed 's/pods\///') yarn run seed
 
 kube-create:
-	make kube-build && \
-	make kube-create-configmaps; \
-	make kube-create-secrets; \
+	docker build -t ddn . && \
 	kubectl apply -f ./k8s
 
-kube-build:
-	docker build -t ddn .
-
-kube-create-configmaps:
-	# kubectl create configmap nginx-conf --from-file ./nginx/prod/nginx.conf -o yaml | kubectl replace -f -
-
-kube-create-secrets:
-	# kubectl create secret tls tls-certificate \
-	# 	--key ./nginx/prod/nginx-selfsigned.key \
-	# 	--cert ./nginx/prod/nginx-selfsigned.crt; \
-	# kubectl create secret generic tls-dhparam \
-	# 	--from-file=./nginx/prod/dhparam.pem
-	# kubectl create secret generic nginx-ssl \
-	# 	--from-file=./nginx/prod/nginx-selfsigned.crt \
-	# 	--from-file=./nginx/prod/nginx-selfsigned.key
-
 kube-delete:
-	kubectl delete -f ./k8s; \
-	kubectl delete configmap nginx-conf config; \
-	kubectl delete secrets tls-certificate tls-dhparam
-	# kubectl delete secrets nginx-ssl; \
-
-kube-url:
-	minikube service nginx-service --url
+	kubectl delete -f ./k8s
