+++ nginx/dev/nginx.conf
@@ -10,8 +10,8 @@ http {
   server {
     listen              443;
     ssl                 on;
-    ssl_certificate     /etc/ssl/certs/nginx-selfsigned.crt;
-    ssl_certificate_key /etc/ssl/certs/nginx-selfsigned.key;
+    ssl_certificate     /etc/ssl/certs/tls.crt;
+    ssl_certificate_key /etc/ssl/certs/tls.key;
 
     location / {
       proxy_pass http://app-web;
