--- nginx/prod/dhparam.pem
@@ -1,8 +1,0 @@
------BEGIN DH PARAMETERS-----
-MIIBCAKCAQEArwnVhjnsEWKFvJE4G+GTrH/GqBNRbotcDdU5N4h0EmXUtA5Prc6e
-+c9jAqGMcUpJAAAu2NeLM+XksVFk94kSpyMHLjDmS8DrQOpOyZS182MA/d0JPv14
-dePg/NIAClgdIIJjm0i0mswz5KHdR4xvsMuVc/kVYeydDFGFPF3SmO4a93v2A8su
-In0KhpssH9cG0RmoBUgC/mCsn26k3uEuYjB5CJDk4GRFH7ZqfZVrp2cKlqeQN/p0
-yUnbfBGi/jaKfa3hmfvuO66uJyyWdVoKPnj2ZDJmQlgWzNRq6tJiagvWYPutW71A
-W38G0lugrM/BHkHTUO+ak/cc/gjf0tANiwIBAg==
------END DH PARAMETERS-----
