--- nginx/prod/nginx-selfsigned.crt
@@ -1,18 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIC+DCCAeACCQD1H5BxVk4npzANBgkqhkiG9w0BAQsFADA+MQswCQYDVQQGEwJU
-VzEPMA0GA1UECAwGVGFpd2FuMQ8wDQYDVQQHDAZUYWlwZWkxDTALBgNVBAoMBG5v
-bmUwHhcNMTgwMjE2MTAwMDQ2WhcNMTkwMjE2MTAwMDQ2WjA+MQswCQYDVQQGEwJU
-VzEPMA0GA1UECAwGVGFpd2FuMQ8wDQYDVQQHDAZUYWlwZWkxDTALBgNVBAoMBG5v
-bmUwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDjYO3rAijFrtuJENgU
-xIV5JBNKTOvCb32EiXprZelpshjcT1b0qfnYmlfp7BolWd+qwWWDmT0AHYWLTpNG
-u0RIzQXWsO0FXvW5eXi/TQnPpInE1pjQ5wndR6dtVMzF/sBqCqEhCOTJpK0afWae
-Ww3BRWm5pvGx2cURxzpUZdXvsMv9Y9mtVPlikOngBPciEBzBQApUd3aGFdnXKh+X
-mJsC8cJZYnZvi2uuJa6Lw5G1bJZfagVmWbKb1H2BqhKLgS1jeBJqDZMDtDggM+H1
-/xGXBlKVHblGibr2gD9vOdTGFb44mkVP3SvObMLtQZRMlhtzWX/A5ZDwmH+z+HvA
-WvLRAgMBAAEwDQYJKoZIhvcNAQELBQADggEBALiFV4wuLy396c0jlLCKou6I3/rY
-0JFIOEyRZZd37CfZk17Tyg4Q4hdvQ0+n5/EPtT2gfS6ZgxfqG9y8pdD8GEPqoOF5
-lMgifjX5hZci5u+aMCvrnRVZNQGryXgc8DstyxQZTk0/wKJXl345XS6n2eqVIoDP
-3vVGByG4QafLJSjMGdTcZaQxrDqSs1E9N3gBdib2l7/qNq5Q7atpiZW4k4Pyjc76
-MZlqH9tGz9mELfCB57AR9/IieerYj2tkxxy8ijMD55o0EaMeoAnFJAPo1egVjEYv
-af/4AHEMFmjqG6dAjV9/PDUrcao0yErVZ3H69A9ObTrSnccQoW2SIO6GFdI=
------END CERTIFICATE-----
