--- nginx/prod/nginx-selfsigned.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDjYO3rAijFrtuJ
-ENgUxIV5JBNKTOvCb32EiXprZelpshjcT1b0qfnYmlfp7BolWd+qwWWDmT0AHYWL
-TpNGu0RIzQXWsO0FXvW5eXi/TQnPpInE1pjQ5wndR6dtVMzF/sBqCqEhCOTJpK0a
-fWaeWw3BRWm5pvGx2cURxzpUZdXvsMv9Y9mtVPlikOngBPciEBzBQApUd3aGFdnX
-Kh+XmJsC8cJZYnZvi2uuJa6Lw5G1bJZfagVmWbKb1H2BqhKLgS1jeBJqDZMDtDgg
-M+H1/xGXBlKVHblGibr2gD9vOdTGFb44mkVP3SvObMLtQZRMlhtzWX/A5ZDwmH+z
-+HvAWvLRAgMBAAECggEAcFhLFYu6k3oC9KVC7nfvLty5P0PkVOx7SwoWV4xCGLwr
-dmcpIrf9P517BJRokr3mOltWCgCj2DO98QfA0vjiE0phsSM36ShzJtIy6+pnh5XX
-kE8rsEQHZrLpsaZtiwKFFZ/s3wKi8FKSSP1ev61jjXP7egRXdTWRHmielGbAnq2A
-n5A8QDWEJe9schytziOqWQK1sBmuNk79Gor6xKLCrMc28b12JeIIzflJvCrYOTgD
-6t+OpPANpdycEz7PnTL+Q/W61hvEQzgOHaS5BN+fcfOC3cUFi9+Jl68ibNL48A6E
-czx/vvqBgNc6Sxe5pTcDU7a6OiQrMRS8ElMmXZ69MQKBgQDzle2+yySznemV2W7x
-f92APlGgeO+yhajDVeRjeb5U1uy5OPREymItLbl8RUWFieZjVPy/igN79s8T8nBT
-p1M1jfwvcb9ee/Fens3Fqy88OCjsmqPeO569TWnLUD1id1xDCMZOrfdx7RYRz0Yz
-QcgIJ4tUVFRBnzJ6llcjuwFx7QKBgQDu94wFhLQXmiu9mrmpjisc3aaVSTamIB6C
-GYPdVxI2rhqOwIduK3bwByVWn3ns7O56C1MJgNw+tZLxmND4ZAE5oH+y502X8Fhr
-DAmZQ3ZUHvA1Nnl7R5hGpU+Un1W0p0I/rmjGl7Jhrfnwarld+60EQuUe7xES8Fi8
-pmp2tps39QKBgQC8y/xV4UY1IbQmvTzMWfrop2TFKHdAu3Au1qgKohXv1BXvCkHp
-/6NQuRwHQfm8o0G6w+iS3PoOXyHpqAOAEzTfUYmhjFxzc0WFvThoeWJHHpA5szJH
-0ZYby6CwQSUS6MjN/EMYQKaF2uk8LJeQqu/Lwjqvxhz4L9gEQ/qOayeXkQKBgBVX
-fGTdxiu5dCCZ2ez/qnAIh6EbicrqYW8jEjjJVvIVOH/5R98Z+KRVLwt+cKgtPNsf
-ecoGfSUImFgFWtxkKgvYfq8eozpnl63vxEew07eNj8xg0halfT8J8oKGfGQG8A65
-59ecoqzAmerTzTxMhtYzlzzdAvxbBMHMZwO1KUWlAoGBAMpXF97fyjqXZoeMgZf4
-sLe/CMISNto2Flzz8JDjRE48aYlDJjLFQBQxo3pKjI50rOmAUINtPKHGvaXyp8Cj
-3wSsQNt2ETq7FAW5i3Y2+c34vbFYKfNbD7pxYVeeT6Qt4a+4MvdMTOlu0PT4CfrH
-WiHcutcnVoXshKVt5xnzd5yG
------END PRIVATE KEY-----
