--- nginx/prod/nginx.conf
@@ -1,22 +1,0 @@
-worker_processes 4;
-
-events { worker_connections 1024; }
-
-http {
-  upstream app-web {
-    ip_hash;
-    server app-service.default:8080;
-  }
-
-  server {
-    listen              80;
-    listen              443;
-    # ssl                 on;
-    ssl_certificate     /etc/ssl/certs/nginx-selfsigned.crt;
-    ssl_certificate_key /etc/ssl/certs/nginx-selfsigned.key;
-
-    location / {
-      proxy_pass http://app-web;
-    }
-  }
-}
